Switch to consistent `policy` field instead of `policyDocument`
