--- src/modules/aws_acm/rpcs/common.ts
@@ -1,7 +1,0 @@
-export const safeParse = (json: string) => {
-  try {
-    return JSON.parse(json);
-  } catch (_) {
-    return undefined;
-  }
-};
