+++ src/modules/aws_acm/rpcs/request.ts
@@ -12,12 +12,12 @@ import { createWaiter, WaiterState } from '@aws-sdk/util-waiter';
 
 import { AwsAcmModule } from '..';
 import { AWS, paginateBuilder } from '../../../services/aws_macros';
+import { safeParse } from '../../../services/common';
 import { awsRoute53Module } from '../../aws_route53';
 import { HostedZone, ResourceRecordSet } from '../../aws_route53/entity';
 import { modules } from '../../iasql_functions/iasql';
 import { Context, RpcBase, RpcInput, RpcResponseObject } from '../../interfaces';
 import { Certificate } from '../entity';
-import { safeParse } from './common';
 
 enum ValidationMethod {
   DNS = 'DNS',
