+++ src/modules/aws_s3/entity/bucket.ts
@@ -40,7 +40,7 @@ export class Bucket {
     type: 'json',
     nullable: true,
   })
-  policyDocument?: Policy;
+  policy?: Policy;
 
   /**
    * @public
