+++ src/modules/aws_vpc/mappers/endpoint_gateway.ts
@@ -10,7 +10,7 @@ import {
 import { AwsVpcModule } from '..';
 import { policiesAreSame } from '../../../services/aws-diff';
 import { AWS, crudBuilderFormat } from '../../../services/aws_macros';
-import { isString } from '../../../services/common';
+import { safeParse } from '../../../services/common';
 import { Context, Crud2, MapperBase } from '../../interfaces';
 import { EndpointGateway } from '../entity';
 import {
@@ -31,7 +31,7 @@ export class EndpointGatewayMapper extends MapperBase<EndpointGateway> {
     Object.is(a.service, b.service) &&
     // the policy document is stringified json
     // we are trusting aws won't change it from under us
-    policiesAreSame(a.policyDocument, b.policyDocument) &&
+    policiesAreSame(a.policy, b.policy) &&
     Object.is(a.state, b.state) &&
     Object.is(a.vpc?.vpcId, b.vpc?.vpcId) &&
     Object.is(a.routeTableIds?.length, b.routeTableIds?.length) &&
@@ -50,7 +50,7 @@ export class EndpointGatewayMapper extends MapperBase<EndpointGateway> {
       (await this.module.vpc.db.read(ctx, this.module.vpc.generateId({ vpcId: eg.VpcId ?? '', region }))) ??
       (await this.module.vpc.cloud.read(ctx, this.module.vpc.generateId({ vpcId: eg.VpcId ?? '', region })));
     if (!out.vpc) return undefined;
-    out.policyDocument = eg.PolicyDocument ? JSON.parse(eg.PolicyDocument) : null;
+    out.policy = eg.PolicyDocument ? safeParse(eg.PolicyDocument) : null;
     out.state = eg.State;
     out.routeTableIds = eg.RouteTableIds;
     if (eg.Tags?.length) {
@@ -90,8 +90,8 @@ export class EndpointGatewayMapper extends MapperBase<EndpointGateway> {
           ServiceName: service,
           VpcId: e.vpc?.vpcId,
         };
-        if (e.policyDocument) {
-          input.PolicyDocument = JSON.stringify(e.policyDocument);
+        if (e.policy) {
+          input.PolicyDocument = JSON.stringify(e.policy);
         }
         if (e.routeTableIds?.length) {
           input.RouteTableIds = e.routeTableIds;
@@ -158,12 +158,12 @@ export class EndpointGatewayMapper extends MapperBase<EndpointGateway> {
         const isUpdate = this.module.endpointGateway.cloud.updateOrReplace(cloudRecord, e) === 'update';
         if (isUpdate) {
           let update = false;
-          if (!policiesAreSame(cloudRecord.policyDocument, e.policyDocument)) {
+          if (!policiesAreSame(cloudRecord.policyDocument, e.policy)) {
             // VPC endpoint policy document update
             const input: ModifyVpcEndpointCommandInput = {
               VpcEndpointId: e.vpcEndpointId,
-              PolicyDocument: JSON.stringify(e.policyDocument),
-              ResetPolicy: !e.policyDocument,
+              PolicyDocument: JSON.stringify(e.policy),
+              ResetPolicy: !e.policy,
             };
             await modifyVpcEndpoint(client.ec2client, input);
             update = true;
