+++ test/modules/aws-vpc-endpoint-interface-integration.ts
@@ -300,7 +300,7 @@ describe('VPC Endpoint interface Integration Testing', () => {
     query(
       `
     UPDATE endpoint_interface
-    SET policy_document = '${JSON.stringify(testPolicy)}'
+    SET policy = '${JSON.stringify(testPolicy)}'
     WHERE tags ->> 'Name' = '${lambdaVpcEndpoint}';
   `,
       undefined,
@@ -327,7 +327,7 @@ describe('VPC Endpoint interface Integration Testing', () => {
       `
     SELECT * FROM endpoint_interface WHERE tags ->> 'Name' = '${lambdaVpcEndpoint}';
   `,
-      (res: any) => expect(res[0]['policy_document']).toEqual(testPolicy),
+      (res: any) => expect(res[0]['policy']).toEqual(testPolicy),
     ),
   );
